fix(minor fixes), doc(add): added new parameters to swagger

+++ src/main/java/com/weather/api/emibeanatte/security/repository/RolRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RolRepository extends JpaRepository<Role, Integer> {
+public interface RolRepository extends JpaRepository<Role, Long> {
 
     Optional<Role> findByRoleName(RoleName roleName);
